requirements.txt
tensorflow~=2.19.0
deepface~=0.0.93
tf-keras~=2.19.0
google-generativeai~=0.8.4
